# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "{{.Description}}",
        "title": "{{.Title}}",
        "contact": {},
        "license": {},
        "version": "{{.Version}}"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
        "/rates": {
            "get": {
                "description": "get all rates available in the system",
                "tags": [
                    "rates"
                ],
                "summary": "Get all rates",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.RateCollection"
                        }
                    }
                }
            },
            "post": {
                "description": "upload rates using json payload",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "rates"
                ],
                "summary": "Upload rates",
                "parameters": [
                    {
                        "description": "Add rates",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.RateCollection"
                        }
                    }
                ],
                "responses": {
                    "201": {}
                }
            }
        },
        "/rates/search": {
            "get": {
                "description": "search rate by start and end date.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "rates"
                ],
                "summary": "Search rate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "date/time as ISO-8601 with timezone",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "date/time as ISO-8601 with timezone",
                        "name": "end",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Rate": {
            "type": "object",
            "properties": {
                "days": {
                    "type": "string",
                    "example": "mon,tues,thurs"
                },
                "price": {
                    "type": "integer",
                    "example": 1500
                },
                "times": {
                    "type": "string",
                    "example": "0900-2100"
                },
                "tz": {
                    "type": "string",
                    "example": "America/Chicago"
                }
            }
        },
        "model.RateCollection": {
            "type": "object",
            "properties": {
                "rates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Rate"
                    }
                }
            }
        }
    }
}